+++ src/services/generate-default-data.ts
@@ -200,7 +200,11 @@ function parseFormSubmissionModel(
   return parseUnknownAsRecord(value, (record) => {
     return elements.reduce<FormSubmissionModel>((model, element) => {
       switch (element.type) {
-        case 'section':
+        case 'section': {
+          const partialModel = parseFormSubmissionModel(element.elements, model)
+          Object.assign(model, partialModel)
+          break
+        }
         case 'page': {
           const partialModel = parseFormSubmissionModel(element.elements, {})
           Object.assign(model, partialModel)